% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridiculousRcade_package.R
\docType{package}
\encoding{UTF-8}
\name{ridiculousRcade}
\alias{ridiculousRcade}
\title{Play Games in your Console}
\description{
This package contains two functions, tic_tac_toe() and rock_paper_scissors()
}
\author{
Carolin Bauer < bauer.carolin@campus.lmu.de >

Stefanie Halmburger <S.Halmburger@campus.lmu.de>
}
