% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock_paper_scissors.R
\name{rock_paper_scissors}
\alias{rock_paper_scissors}
\title{Play rock paper scissors}
\usage{
rock_paper_scissors(choice)
}
\arguments{
\item{choice}{A string. Choices are "rock", "paper", or "scissors"}
}
\value{
A character string with result.
}
\description{
Play rock paper scissors
}
\examples{
rock_paper_scissors(choice = "paper")
}
